fix bug

+++ app/Http/Livewire/Frontend/Jobs/ShowForm.php
@@ -533,18 +533,19 @@ class ShowForm extends Component
                 $remote_id =  $this->job->remotejob->remote_job_id;
                 $pid_shell = sprintf("qsig -s SIGKILL %s" ,$remote_id );
                 $server = sshservers::find($this->server_id);
-                
-                $c = new Connection($server->server_name, $server->host.":".$server->port, $server->username,["password"=>$server->password]);
-                $c->run([
-                    $pid_shell
-                ], function($line) use($c)
-                {
+                if ($server) {
+                    $c = new Connection($server->server_name, $server->host.":".$server->port, $server->username,["password"=>$server->password]);
                     $c->run([
-                        sprintf('kill %s', $line)
-                    ], function($line2) {
-                    });
+                        $pid_shell
+                    ], function($line) use($c)
+                    {
+                        $c->run([
+                            sprintf('kill %s', $line)
+                        ], function($line2) {
+                        });
+                    }
+                    );
                 }
-                );
             }
         } catch(Exception $e) {
             dd($e->getMessage());
@@ -580,7 +581,7 @@ class ShowForm extends Component
     {
         try {
             $this->job = job::find($jobID);
-            if($this->job->progress == 'Completed') {
+            if($this->job->progress == 'Completed' || $this->job->progress == 'Cancelled' ) {
                 $output_name = sprintf("%d_output.zip", $jobID);
                 $server = sshservers::find($server_id);
                 $sftp = new SFTP($server->host, $server->port);
